Remove $CONFIG_FORK branch for buildkite-config checkout
This is useful when triggering builds manually from bk-c, when working from a fork.

Since it's not really a common use-case for most developers, we can remove it.
The recommended way to trigger builds is from a PR.

If you do need to manually trigger builds, you can change the steps yourself.